backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
python-jose[cryptography] 
passlib[bcrypt]
python-multipart
pydantic[email]